+++ src/parser.js
@@ -1,5 +1,5 @@
 // parser.js
-// (c) tearflake, 2024
+// (c) tearflake, 2025
 // MIT License
 
 var Parser = (